#25 -change target framework version to 4.7.2 -add editor only library -add shared library for use by any project -remove code duplication -add automated copy step in UnrealEngine.Runtime.sln via Directory.Build.props -> this removes the need to specify the output directory manually. it also helps when the output path is extended by MSBUILD on its own (the case for netstandart libs) -adding code to load editor library in editor
