+++ Managed/UnrealEngine.Runtime/UnrealEngine.Runtime.sln
@@ -12,7 +12,12 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F98E80EC-0F91-42F0-9E1D-0FE57BDC924A}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
+		Directory.Build.props = Directory.Build.props
 	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "USharp.Editor", "USharp.Editor\USharp.Editor.csproj", "{92B3DB21-34C4-43CD-87A5-D62A0D13B1BE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "USharp.Shared", "USharp.Shared\USharp.Shared.csproj", "{728A3B8C-A410-4DEC-8BC7-47D81DA2B979}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,6 +36,14 @@ Global
 		{C284889D-AA3A-4EC1-B4EA-FAA4A3A92C33}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C284889D-AA3A-4EC1-B4EA-FAA4A3A92C33}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C284889D-AA3A-4EC1-B4EA-FAA4A3A92C33}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{92B3DB21-34C4-43CD-87A5-D62A0D13B1BE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{92B3DB21-34C4-43CD-87A5-D62A0D13B1BE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{92B3DB21-34C4-43CD-87A5-D62A0D13B1BE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{92B3DB21-34C4-43CD-87A5-D62A0D13B1BE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{728A3B8C-A410-4DEC-8BC7-47D81DA2B979}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{728A3B8C-A410-4DEC-8BC7-47D81DA2B979}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{728A3B8C-A410-4DEC-8BC7-47D81DA2B979}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{728A3B8C-A410-4DEC-8BC7-47D81DA2B979}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
